+++ src/Exception/RetryUploadFlowException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace InstagramAPI\Exception;
+
+class RetryUploadFlowException extends RequestException
+{
+}